Created register, alu, and immediate generator
Created the register, alu, and immediate generator modules since they are the most needed modules for instructions, especially the register and alu module. The alu module specifically is only created to the standard of TinyRV1 where it doesn't include most of the operations. TinyRV2 operations will be added in the future once TinyRV1 is fully completed.